Refactorización para recibir y enviar el contenido de los documentos en bytes en vez de base64

+++ src/main/java/com/fjld/secure_storage/dto/DocumentResponseDTO.java
@@ -12,7 +12,7 @@ public class DocumentResponseDTO {
     private String uuid;
     private String name;
     private String description;
-    private Integer size;
+    private Long size;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
